+++ UM_LOGISTIC_V1/Client/Controllers/Cooperation/CooperationController.js
@@ -44,7 +44,7 @@
 		.success(response => {
 		    $scope.isPartLoading = false;
 		    $scope.isLoading = false;
-		    if (response.Success) {
+		    if (response && response.Success == true) {
 		        for (let i = 0; i < response.Result.length; i++) {
 		            $scope.cooperations.push({
 		                id: FormHelper.getFormValue(response.Result[i].Id),
@@ -73,7 +73,7 @@
 		.success(response => {
 		    $scope.isLoading = false;
 		    $scope.filterLoading = false;
-		    if (response.Success) {
+		    if (response && response.Success == true) {
 		        if (page == 0) {
 		            $scope.cooperations = [];
 		        }
@@ -114,7 +114,7 @@
         let type = false;
         ApplicationPictureService.getApplicationPictures(id, type)
 		.success(response => {
-		    if (response.Success && response.Result && response.Result.length > 0) {
+		    if (response && response.Success && response.Result && response.Result.length > 0) {
 		        $scope.pictures[id] = response.Result[0];
 		    }
 		    else {
@@ -140,7 +140,7 @@
         request.TypeId = 3;
         ClientTaskService.createApplicationTask(request)
         .success(response => {
-            if (response.Success) {
+            if (response && response.Success == true) {
                 NotificationService.success(moduleConstants.callFeedbackAccepted);
                 $scope.isLoading = false;
             }
@@ -160,7 +160,6 @@
         let filter = $scope.filter;
         let stringFilter = $scope.generateFilterString(filter);
         $scope.currentPage = 0;
-        //$scope.isLoading = true;
         $scope.filterLoading = true;
         $scope.listFilteredCooperations(stringFilter, $scope.currentPage, $scope.currentCount);
     }
@@ -289,10 +288,10 @@
             UserId: userId
         };
         ApplicationTrashService.createApplicationTrash(request).success(response => {
-            if (response.Success == false) {
+            if (response && response.Success == false) {
                 NotificationService.error(response.Error != null ? JSON.stringify(response.Error) : moduleConstants.internalErrorCaption);
             }
-            if (response.Success == true) {
+            if (response && response.Success == true) {
                 let isTrashAdded = SessionService.addTrashElement(id, false, title);
                 if (isTrashAdded == true) {
                     $rootScope.$broadcast("trashElementAdded", null);
@@ -318,7 +317,7 @@
         }
         ApplicationTrashService.removeTrashElement(id, false)
         .success(response => {
-            if (response.Success) {
+            if (response && response.Success == true) {
                 $scope.isLoading = false;
                 SessionService.removeShopTrashElement(id, false);
                 $rootScope.$broadcast("trashElementRemoved", null);
